♻️ extract tool management logic from wizard and clean defaults
- Remove default invoke_subagent tool from toolRegistry config,
  starting with an empty array for cleaner initial setup
- Extract single-agent tool/KB/MCP/sub-agent action handlers out of
  the main wizard component to reduce complexity
- Simplify derived data section to be architecture-agnostic
- Fix indentation for evaluatorConfig supportedModels block
